django_app/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.9
django-bootstrap-v5==1.0.11
django-environ==0.11.2
django-leaflet==0.29.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.1
soupsieve==2.5
sqlparse==0.4.4
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
